refactor UI for dummy export menu
